+++ include/aspas_merge_mic.tcc
@@ -1028,16 +1028,16 @@ in_register_merge(T& v0, T& v1)
 
 template <typename T>
 typename std::enable_if<std::is_same<T, int>::value>::type
-merge(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output)
+merge(T* inputA, uint64_t sizeA, T* inputB, uint64_t sizeB, T* output)
 {
     __m512i vec0 = _mm512_set1_epi32(1);
     __m512i vec1 = _mm512_set1_epi32(1);
     uint8_t stride = (uint8_t)simd_width::AVX512_INT;
-    uint32_t i0=0;
-    uint32_t i1=0;
-    uint32_t iout=0;
+    uint64_t i0=0;
+    uint64_t i1=0;
+    uint64_t iout=0;
     int buffer[stride];
-    uint32_t i3=0;
+    uint64_t i3=0;
 
     if(sizeA >= stride && sizeB >= stride)
     {
@@ -1218,7 +1218,7 @@ merge(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output)
 
 template <typename T>
 typename std::enable_if<std::is_same<T, int>::value>::type
-merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, int *ptrA, int *ptrB, int *output_ptr)
+merge_key(T* inputA, uint64_t sizeA, T* inputB, uint64_t sizeB, T* output, int *ptrA, int *ptrB, int *output_ptr)
 {
     __m512i vec0 = _mm512_set1_epi32(1);
     __m512i vec1 = _mm512_set1_epi32(1);
@@ -1226,12 +1226,12 @@ merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, int *
     __m512i p0 = _mm512_set1_epi32(1);
     __m512i p1 = _mm512_set1_epi32(1);
     uint8_t stride = (uint8_t)simd_width::AVX512_INT;
-    uint32_t i0=0;
-    uint32_t i1=0;
-    uint32_t iout=0;
+    uint64_t i0=0;
+    uint64_t i1=0;
+    uint64_t iout=0;
     int buffer[stride];
     int buffer_ptr[stride];
-    uint32_t i3=0;
+    uint64_t i3=0;
 
     if(sizeA >= stride && sizeB >= stride)
     {
@@ -1451,7 +1451,7 @@ merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, int *
 
 template <typename T>
 typename std::enable_if<std::is_same<T, int>::value>::type
-merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, long *ptrA, long *ptrB, long *output_ptr)
+merge_key(T* inputA, uint64_t sizeA, T* inputB, uint64_t sizeB, T* output, long *ptrA, long *ptrB, long *output_ptr)
 {
     __m512i vec0 = _mm512_set1_epi32(1);
     __m512i vec1 = _mm512_set1_epi32(1);
@@ -1461,12 +1461,12 @@ merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, long 
     __m512i ph0 = _mm512_set1_epi64(1);
     __m512i ph1 = _mm512_set1_epi64(1);
     uint8_t stride = (uint8_t)simd_width::AVX512_INT;
-    uint32_t i0=0;
-    uint32_t i1=0;
-    uint32_t iout=0;
+    uint64_t i0=0;
+    uint64_t i1=0;
+    uint64_t iout=0;
     int buffer[stride];
     long buffer_ptr[stride];
-    uint32_t i3=0;
+    uint64_t i3=0;
 
     if(sizeA >= stride && sizeB >= stride)
     {
@@ -1707,7 +1707,7 @@ merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, long 
 }
 template <typename T>
 typename std::enable_if<std::is_same<T, float>::value>::type
-merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, int *ptrA, int *ptrB, int *output_ptr)
+merge_key(T* inputA, uint64_t sizeA, T* inputB, uint64_t sizeB, T* output, int *ptrA, int *ptrB, int *output_ptr)
 {
     __m512 vec0 = _mm512_set1_ps(1);
     __m512 vec1 = _mm512_set1_ps(1);
@@ -1715,12 +1715,12 @@ merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, int *
     __m512i p0 = _mm512_set1_epi32(1);
     __m512i p1 = _mm512_set1_epi32(1);
     uint8_t stride = (uint8_t)simd_width::AVX512_FLOAT;
-    uint32_t i0=0;
-    uint32_t i1=0;
-    uint32_t iout=0;
+    uint64_t i0=0;
+    uint64_t i1=0;
+    uint64_t iout=0;
     float buffer[stride];
     int buffer_ptr[stride];
-    uint32_t i3=0;
+    uint64_t i3=0;
 
     if(sizeA >= stride && sizeB >= stride)
     {
@@ -1940,7 +1940,7 @@ merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, int *
 
 template <typename T>
 typename std::enable_if<std::is_same<T, float>::value>::type
-merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, long *ptrA, long *ptrB, long *output_ptr)
+merge_key(T* inputA, uint64_t sizeA, T* inputB, uint64_t sizeB, T* output, long *ptrA, long *ptrB, long *output_ptr)
 {
     __m512 vec0 = _mm512_set1_ps(1);
     __m512 vec1 = _mm512_set1_ps(1);
@@ -1950,12 +1950,12 @@ merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, long 
     __m512i ph0 = _mm512_set1_epi64(1);
     __m512i ph1 = _mm512_set1_epi64(1);
     uint8_t stride = (uint8_t)simd_width::AVX512_FLOAT;
-    uint32_t i0=0;
-    uint32_t i1=0;
-    uint32_t iout=0;
+    uint64_t i0=0;
+    uint64_t i1=0;
+    uint64_t iout=0;
     float buffer[stride];
     long buffer_ptr[stride];
-    uint32_t i3=0;
+    uint64_t i3=0;
 
     if(sizeA >= stride && sizeB >= stride)
     {
@@ -2197,7 +2197,7 @@ merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, long 
 
 template <typename T>
 typename std::enable_if<std::is_same<T, double>::value>::type
-merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, int *ptrA, int *ptrB, int *output_ptr)
+merge_key(T* inputA, uint64_t sizeA, T* inputB, uint64_t sizeB, T* output, int *ptrA, int *ptrB, int *output_ptr)
 {
     __m512d vec0 = _mm512_set1_pd(1);
     __m512d vec1 = _mm512_set1_pd(1);
@@ -2206,12 +2206,12 @@ merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, int *
     __m512i p1 = _mm512_set1_epi32(1);
     __mmask16 mp = 0x00ff;
     uint8_t stride = (uint8_t)simd_width::AVX512_DOUBLE;
-    uint32_t i0=0;
-    uint32_t i1=0;
-    uint32_t iout=0;
+    uint64_t i0=0;
+    uint64_t i1=0;
+    uint64_t iout=0;
     double buffer[stride];
     int buffer_ptr[stride];
-    uint32_t i3=0;
+    uint64_t i3=0;
 
     if(sizeA >= stride && sizeB >= stride)
     {
@@ -2451,7 +2451,7 @@ merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, int *
 
 template <typename T>
 typename std::enable_if<std::is_same<T, double>::value>::type
-merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, long *ptrA, long *ptrB, long *output_ptr)
+merge_key(T* inputA, uint64_t sizeA, T* inputB, uint64_t sizeB, T* output, long *ptrA, long *ptrB, long *output_ptr)
 {
     __m512d vec0 = _mm512_set1_pd(1);
     __m512d vec1 = _mm512_set1_pd(1);
@@ -2459,12 +2459,12 @@ merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, long 
     __m512i p0 = _mm512_set1_epi64(1);
     __m512i p1 = _mm512_set1_epi64(1);
     uint8_t stride = (uint8_t)simd_width::AVX512_DOUBLE;
-    uint32_t i0=0;
-    uint32_t i1=0;
-    uint32_t iout=0;
+    uint64_t i0=0;
+    uint64_t i1=0;
+    uint64_t iout=0;
     double buffer[stride];
     long buffer_ptr[stride];
-    uint32_t i3=0;
+    uint64_t i3=0;
 
     if(sizeA >= stride && sizeB >= stride)
     {
@@ -2703,16 +2703,16 @@ merge_key(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output, long 
 }
 template <typename T>
 typename std::enable_if<std::is_same<T, float>::value>::type
-merge(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output)
+merge(T* inputA, uint64_t sizeA, T* inputB, uint64_t sizeB, T* output)
 {
     __m512 vec0 = _mm512_set1_ps(1.f);
     __m512 vec1 = _mm512_set1_ps(1.f);
     uint8_t stride = (uint8_t)simd_width::AVX512_FLOAT;
-    uint32_t i0=0;
-    uint32_t i1=0;
-    uint32_t iout=0;
+    uint64_t i0=0;
+    uint64_t i1=0;
+    uint64_t iout=0;
     float buffer[stride];
-    uint32_t i3=0;
+    uint64_t i3=0;
 
     if(sizeA >= stride && sizeB >= stride)
     {
@@ -2893,17 +2893,17 @@ merge(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output)
 
 template <typename T>
 typename std::enable_if<std::is_same<T, double>::value>::type
-merge(T* inputA, uint32_t sizeA, T* inputB, uint32_t sizeB, T* output)
+merge(T* inputA, uint64_t sizeA, T* inputB, uint64_t sizeB, T* output)
 {
     __m512d vec0 = _mm512_set1_pd(1.0);
     __m512d vec1 = _mm512_set1_pd(1.0);
 
     uint8_t stride = (uint8_t)simd_width::AVX512_DOUBLE;
-    uint32_t i0=0;
-    uint32_t i1=0;
-    uint32_t iout=0;
+    uint64_t i0=0;
+    uint64_t i1=0;
+    uint64_t iout=0;
     double buffer[stride];
-    uint32_t i3=0;
+    uint64_t i3=0;
 
     if(sizeA>=stride && sizeB>=stride)
     {
