+++ include/sorter_avx.tcc
@@ -43,7 +43,7 @@ namespace internal
  *
  */
 template <typename T>
-void swap(T* a, uint32_t i, uint32_t j) 
+void swap(T* a, uint64_t i, uint64_t j) 
 {
     if(a[i] > a[j]) 
     {
@@ -54,7 +54,7 @@ void swap(T* a, uint32_t i, uint32_t j)
 }
 
 template <typename T>
-void swap_key(T* a, int *ptr, uint32_t i, uint32_t j) 
+void swap_key(T* a, int *ptr, uint64_t i, uint64_t j) 
 {
     if(a[i] > a[j]) 
     {
@@ -69,7 +69,7 @@ void swap_key(T* a, int *ptr, uint32_t i, uint32_t j)
 }
 
 template <typename T>
-void swap_key(T* a, long *ptr, uint32_t i, uint32_t j) 
+void swap_key(T* a, long *ptr, uint64_t i, uint64_t j) 
 {
     if(a[i] > a[j]) 
     {
@@ -1828,9 +1828,9 @@ in_register_transpose_key(T& v0, T& v1, T& v2, T& v3,
 
 template <typename T>
 typename std::enable_if<std::is_same<T, int>::value>::type
-sorter(T*& orig, uint32_t size)
+sorter(T*& orig, uint64_t size)
 {
-    uint32_t i, j;
+    uint64_t i, j;
     __m256i vec0;
     __m256i vec1;
     __m256i vec2;
@@ -1902,9 +1902,9 @@ sorter(T*& orig, uint32_t size)
 
 template <typename T> // int key int index
 typename std::enable_if<std::is_same<T, int>::value>::type
-sorter_key(T*& orig, int *&ptr, uint32_t size)
+sorter_key(T*& orig, int *&ptr, uint64_t size)
 {
-    uint32_t i, j;
+    uint64_t i, j;
     __m256i vec0;
     __m256i vec1;
     __m256i vec2;
@@ -2007,9 +2007,9 @@ sorter_key(T*& orig, int *&ptr, uint32_t size)
 
 template <typename T> // int key long index
 typename std::enable_if<std::is_same<T, int>::value>::type
-sorter_key(T*& orig, long *&ptr, uint32_t size)
+sorter_key(T*& orig, long *&ptr, uint64_t size)
 {
-    uint32_t i, j;
+    uint64_t i, j;
     __m256i vec0;
     __m256i vec1;
     __m256i vec2;
@@ -2141,9 +2141,9 @@ sorter_key(T*& orig, long *&ptr, uint32_t size)
 
 template <typename T>
 typename std::enable_if<std::is_same<T, float>::value>::type
-sorter(T*& orig, uint32_t size)
+sorter(T*& orig, uint64_t size)
 {
-    uint32_t i, j;
+    uint64_t i, j;
     __m256 vec0;
     __m256 vec1;
     __m256 vec2;
@@ -2215,9 +2215,9 @@ sorter(T*& orig, uint32_t size)
 
 template <typename T> // float key int index
 typename std::enable_if<std::is_same<T, float>::value>::type
-sorter_key(T*& orig, int *&ptr, uint32_t size)
+sorter_key(T*& orig, int *&ptr, uint64_t size)
 {
-    uint32_t i, j;
+    uint64_t i, j;
     __m256 vec0;
     __m256 vec1;
     __m256 vec2;
@@ -2320,9 +2320,9 @@ sorter_key(T*& orig, int *&ptr, uint32_t size)
 
 template <typename T> // float key long index
 typename std::enable_if<std::is_same<T, float>::value>::type
-sorter_key(T*& orig, long *&ptr, uint32_t size)
+sorter_key(T*& orig, long *&ptr, uint64_t size)
 {
-    uint32_t i, j;
+    uint64_t i, j;
     __m256 vec0;
     __m256 vec1;
     __m256 vec2;
@@ -2455,9 +2455,9 @@ sorter_key(T*& orig, long *&ptr, uint32_t size)
 
 template <typename T>
 typename std::enable_if<std::is_same<T, double>::value>::type
-sorter(T*& orig, uint32_t size)
+sorter(T*& orig, uint64_t size)
 {
-    uint32_t i, j;
+    uint64_t i, j;
     __m256d vec0;
     __m256d vec1;
     __m256d vec2;
@@ -2501,9 +2501,9 @@ sorter(T*& orig, uint32_t size)
 
 template <typename T> // double key int index
 typename std::enable_if<std::is_same<T, double>::value>::type
-sorter_key(T*& orig, int *&ptr,  uint32_t size)
+sorter_key(T*& orig, int *&ptr,  uint64_t size)
 {
-    uint32_t i, j;
+    uint64_t i, j;
     __m256d vec0;
     __m256d vec1;
     __m256d vec2;
@@ -2556,9 +2556,9 @@ sorter_key(T*& orig, int *&ptr,  uint32_t size)
 
 template <typename T> // double key long index
 typename std::enable_if<std::is_same<T, double>::value>::type
-sorter_key(T*& orig, long *&ptr, uint32_t size)
+sorter_key(T*& orig, long *&ptr, uint64_t size)
 {
-    uint32_t i, j;
+    uint64_t i, j;
     __m256d vec0;
     __m256d vec1;
     __m256d vec2;
